be/requirements.txt
aiofiles==0.7.0
aiomysql==0.2.0
aiosmtplib==1.1.4
aiosqlite==0.16.0
attrs==21.2.0
bcrypt==4.0.1
grpcio==1.59.0
httpx==0.27.2
Jinja2==3.1.2
loguru==0.7.2
lxml==4.9.3
MarkupSafe==2.1.2
Pillow==10.0.1
protobuf==3.20.3
PyJWT==2.6.0
PyMySQL==1.0.2
pytz==2024.2
sanic==23.12.1
tortoise-orm==0.20.0
flake8==6.1.0
croniter==2.0.1
freezegun==1.2.1
tomli==2.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
redis==5.0.8
sanic-ext==23.12.0
orjson==3.10.7
motor==3.6.0